Add cleaning steps to ensure CSVs as like the ones used in the paper
After running all experiments and finalizing the paper, we
noticed that the tables used to run experiments are slightly
different from the tables published as supplemental material.
While we would prefer to publish the raw data, to ensure full
reproducility and transparency, we (the paper authors) decided
to include these data cleaning steps here to make sure that the
same files used in our experiments can be recreated.
We will consider removing some of these cleaning steps in a
future release of the dataset.